# code-titan-ems-api-doc.yaml
openapi: 3.0.1
info:
  title: Code Titan's Event Management System
  version: 1.0.0
  description: An API for a web-based Event Management System
servers:
  - url: 'http://127.0.0.1:5000'
paths:
  /api/events:
    summary: View all events
    get:
      summary: View all events
      description: 'This api is responsible for generating all the events available'
      operationId: viewEvents
      security:
        - BearerAuth: []
      
      responses:
        '200':
          description: Events were loaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      events:
                        $ref: '#/components/schemas/Event'
                      
                   
                          
                  error:
                    example :
                      
                      
                  message:
                    type: string
                    example : "success"
        '204':
                    
          description: No event exist 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/newEventCreated'
                
              examples:
                No Result:
                  $ref: '#/components/examples/noResult'

        '401':
          $ref: '#/components/responses/UnauthorizedError'
      tags:
        - View
        
    post:
      summary: create new events
      description: 'This api is responsible for creating a new event'
      operationId: createEvents
      security:
        - BearerAuth: []
        
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/newEvent' 
            examples:
              new Event:
                $ref: '#/components/examples/newEvent'
        
      parameters:
        - in: query
          name: title
          description: The title of the event
          schema:
            type: string
          required: true
        - in: query
          name: start_date
          description: The start date of the event (YYYY-MM-DD)
          schema:
            type: string
          required: true
        - in: query
          name: end_date
          description: The end date of the event (YYYY-MM-DD)
          schema:
            type: string
          required: true
        - in: query
          name: description
          description: Description of the event
          schema:
            type: string
          required: true
        - in: query
          name: venue
          description: The venue of the event 
          schema:
            type: string
          required: true
        - in: query
          name: photo
          description: Promotional Image of event (jpg, jpeg, png)
          schema:
            type: file
          required: true
        - in: query
          name: url
          description: URL to  event
          schema:
            type: string
          required: true
      
      responses:
        '200':
          description: New Event was created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/newEventCreated'
                
              examples:
                new Event:
                  $ref: '#/components/examples/newEventCreated'
              
                
      tags:
        - Create
  /api/events/{uid}:
    get:
      summary: Filter and view all events by user ID
      description: 'This api is responsible for generating all the events for a specific user'
      operationId: viewEventsByUID
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: uid
          description: The unique user ID of events 
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Events were loaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
              examples:
                UID = 1:
                  $ref: '#/components/examples/viewEvent'
                  
        '204':
                    
          description: No event with  UID Matched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/newEventCreated'
                
              examples:
                No Result:
                  $ref: '#/components/examples/noResult'
      tags:
        - View
  /api/events/{event_id}:
    delete:
      summary: Delete event by event ID
      description: 'This api is responsible for deleting events by event id '
      operationId: deleteEventsByID
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: event_id
          description: The unique user ID of events 
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: ""
          content:
            application/json:
              examples:
                UID = 5:
                  $ref: '#/components/examples/Success'
                  
        '204':
          description: "Event ID was not found"
          content:
            application/json:
              examples:
                No Match:
                  $ref: '#/components/examples/failUpdate'  
                
                
          
      tags:
        - Delete
  /api/events/{status}:
    get:
      summary: Filter and view events by events status
      description: 'This api is responsible for viewing events by event status specified'
      operationId: ""
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: status
          description: "Event status to filter by "
          schema:
            type: string
          required: true
          
      responses:
        '200':
          description: Events filtered by event status are loaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
              examples:
                Published:
                  $ref: '#/components/examples/viewEvent'
        '204':
                    
          description: No event with status matched 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/newEventCreated'
                
              examples:
                No Result:
                  $ref: '#/components/examples/noResult'
          
      
      tags:
        - View
  /api/events/search/date/{search_date}:
    get:
      summary: Searches for events taking place on the search date with format yyyy-mm-dd
      description: 'Returns a list of events by date specified'
      parameters:
      - name: "search_date"
        in: "path"
        description: "Date of event"
        required: true
        schema:
          type: "string"
          format: "unicode"
      
      responses:
        '200':
          description: "Date found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
              examples:
                2021-08-05:
                  $ref: '#/components/examples/viewEvent'
        '204':
          description: "No date matched"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
              examples:
                Date not found:
                  $ref: '#/components/examples/noDateFound'
          
          
          
        '400':
          description: Invalid Date format
          
      tags:
        - Search
  /api/events/search/title/{title}:
    get:
      summary: Searches for events based on Title
      description: 'Returns a list of events by title specified'
      parameters:
      - name: title
        in: "path"
        description: "Title of event"
        required: true
        schema:
          type: "string"
      # requestBody:
      #   content:
      #     application/json:
      #       schema:      # Request body contents
      #         type: object
      #         properties:
      #           id:
      #             type: integer
      #           title:
      #             type: string    
                
      #           start_date:
      #               type: datetime
      #           end_date:
      #               type: datetime
      #           description: 
      #               type: string
      #           venue: 
      #               type: string
      #           photo:
      #               type: string
      #           website_url:
      #               type: string
      #           status:
      #               type: string
      #           uid:
      #               type: integer
      #           created:
      #               type: datetime           
              
              # example:   # Sample object
              #   id: 10
              #   name: Jessica Smith            
                
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
              examples:
                Event 1:
                  $ref: '#/components/examples/viewEvent'
          
          
        '204':
          description: "No title found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
              examples:
                No Match:
                  $ref: '#/components/examples/noTitleFound'
          
        
          
      tags:
        - Search
  /api/events/publish/{event_id}:
    put:
      summary: Changes the status of an event to published
      description: This api is responsible for updating event's status to published
      parameters:
      - name: event_id
        in: "path"
        description: "Event ID"
        required: true
        schema:
          type: integer
      # requestBody:
      #   content:
      #       application/json:
      #           schema:      # Request body contents
      #           type: object
      #           properties:
      #               id:
      #               name:
      #               type: string
      #           example:   # Sample object
      #               id: 10
      #               name: Jessica Smith    
      responses:
        '200':
          description: OK
          content:
            application/json:
              examples:
                UID = 1:
                  $ref: '#/components/examples/Success'
        '204':
          description: "Event ID was not found"
          content:
            application/json:
              examples:
                No Match:
                  $ref: '#/components/examples/failUpdate'
      tags:
        - Update
  /api/events/reject/{event_id}:
    put:
      summary: Changes the status of an event to rejected
      description: This api is responsible for updating event's status to rejected
      parameters:
      - name: event_id
        in: "path"
        description: "Event ID"
        required: true
        schema:
          type: integer
      # requestBody:
      #   content:
      #       application/json:
      #           schema:      # Request body contents
      #           type: object
      #           properties:
      #               id:
      #               name:
      #               type: string
      #           example:   # Sample object
      #               id: 10
      #               name: Jessica Smith    
      responses:
        '200':
          description: OK
          content:
            application/json:
              examples:
                UID = 1:
                  $ref: '#/components/examples/Success'
        '204':
          description: "Event ID was not found"
          content:
            application/json:
              examples:
                No Match:
                  $ref: '#/components/examples/failUpdate'  
      tags:
        - Update
components:
  schemas:
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        start_date:
          type: string
          format: "Datetime"
        end_date: 
          type: string
          format: "Datetime"
        description : 
          type: string
        venue : 
          type: string
        photo : 
          type: string
          format: "jpg, jpeg, png"
        url : 
          type: string
        status:
          type: "string"
          description: "Event Statuses"
          enum:
            - "Pending"
            - "Published"
            - "Rejected"
        uid: 
          type: integer
        created_at : 
          type: string
          format: "Datetime"
    newEvent:
      type: object
      properties:
        title:
          type: string
        start_date:
          type: string
          format: "Datetime"
        end_date: 
          type: string
          format: "Datetime"
        description : 
          type: string
        venue : 
          type: string
        photo : 
          type: string
        url : 
          type: string
        status:
          type: "string"
    newEventCreated:
      type: object
      properties:
        data:
          type: object
          properties:
            events:
              $ref: '#/components/schemas/newEvent'
            
        error:
          example :
                      
        message:
          type: string
          example : "success"
  examples:
...
    newEvent:
      value:
        title: "Event 1"
        start_date: "Wed, 04 Aug 2021 00:00:00 GMT"
        end_date: "Wed, 04 Aug 2021 00:00:00 GMT"
        description : " description 1"
        venue : "At home 1"
        photo : "event_pic.png"
        url : "youreventwebsite.com"
        status : "Pending"
    newEventCreated:
      value:
        data:
          title: "Event 1"
          start_date: "Wed, 04 Aug 2021 00:00:00 GMT"
          end_date: "Wed, 04 Aug 2021 00:00:00 GMT"
          description : " description 1"
          venue : "At home 1"
          photo : "event_pic.png"
          url : "youreventwebsite.com"
          status : "Pending"
          uid: 2
          created_at : "Wed, 04 Aug 2021 00:00:00 GMT"
        message: "Event successfully created!"
    Success:
      value:
        error:
        message: "Success"
    failUpdate:
      value:
        error: "No ID Matched"
        message: "No Event Found"
    noResult:
      value:
        data:
          events: []
        error:
        message: "No Event Found"
...
    noDateFound:
      value:
        data:
          events: []
        error:
        message: "No Date Found"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Enter your JWT Bearer Token
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid      
...